Switched from hardcoded types to generated protobuf types

+++ iso8583/types/field.go
@@ -1,7 +1,7 @@
 package types
 
 type field struct {
-	Length int
+	Length int64
 	Value  string
 	Raw    []byte
 }
